requirements.txt
streamlit>=1.28.0
anthropic>=0.7.7
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.0.0
pandas>=2.1.0
python-dotenv>=1.0.0
json5>=0.9.14
plotly>=5.0.0
streamlit-authenticator>=0.2.3 